logger test
